Start adding support for re-assembling instructions
The split of InstructionSet makes it easier to reference Def from
InstructionInstance.  This may be replaced by a reference to Prefixes,
but the same restructuring would be needed.

There is a test suite that 1) assembles an assembly program using gcc,
2) disassembles it using flexdis, and 3) reassembles it using the new
assembler.  The test passes if the output of re-assembly matches the
output from gcc.