Fix a bug with ModR/M computation
Before we had a handy flag in each InstructionInstance telling us if we
needed to add a ModR/M byte at all.  This was removed in favor of a
solution that just figured it out on the fly (to make instruction
creation simpler).

This turned out to be tricky because we needed to know some information
about argument types from the 'Def' to do it properly.  In particular,
for register operands that are encoded in the opcode, that information
is only available in the opcode type.  We were discarding that.  Now we
thread it through everywhere it needs to be (including the
InstructionInstance, which re-computes it from the 'Def').