Improve handling of R/M operands
It turns out that the order of operands only matters somewhat.  What
matters more is that some types of operand MUST be encoded in the R/M
bits, while the other is then encoded in the Reg bits.

This change reorders the operands when required to simplify processing
so that we always put the correct operand into R/M.