Fix the ordering of required prefixes
This might not be entirely correct, but required prefixes are always
treated as one of the legacy prefixes.  That means they come *before*
the REX prefix.

Since they can occur in any order, this is fine.  It might make the
tests complicated if the required prefix is mixed with some other legacy
prefix (e.g., ASO or OSO) and the order is wrong.