Use absolute path for `data/optable.xml`, 2nd try :P
The absolute-path code I implemented in the last commit is broken: the
files listed in a Cabal `data-files` section are only installed
*after* the build completes, whereas the template Haskell of course
needs to run before the build completes. While developing the first
fix, I had done an intermediate build after adding `data/optable.xml`
to the `data-files`, but before changing the template code to call
`getDataFileName`, which meant the `data-files` where available letter
when I updated the template Haskell code; on a fresh build the
`getDataFileName` call always fails.