Change the decoding of IM_SZ values
These are the sIz type in the xml data file.  These are values of ZSize
type, which means that they are always either 16 or 32 bits in the
bitstream.  At run-time, they are sign-extended to the width of the
relevant register.  It is important to decode these to the correct size
so that we can restore them with the re-assembler.