Add a more principled assembler fix for immediates
The core issue is that immediates need to be encoded into different
physical sizes based on a number of prefixes (operand-size-override and
REX.w) and conventions.

The new code tries to be more systematic about it.  Furthermore, it will
fail loudly when a case is not handled so we won't miss it as a weird
decoding error.