Avoid runtime exception in ppImm
When given a negative value, `ppImm` calls `negate` on it and passes the
result to `showHex`. In situations where the negated value is outside of
the range of the type (e.g. > `maxBound`), `negate` can return a
negative value which causes `showHex` to throw a runtime exception.

Example:

    let i = (-127 :: Int8) - 1
    i
    > -128

    i == negate i
    > True

    ppImm i
    > -0x*** Exception: Numeric.showIntAtBase: applied to negative number -128

If we convert to an Integer before negating we can avoid this exception.