assembler: generalize a case
The case was only handling the new set of 64 bit registers, and not the classic
registers.  Now we can match them properly in the assembler.  This affects
encodings of the form "Opcode_reg" where part of the opcode encodes information
about the register.  In particular, we found this while assembling mov
instructions.