Fix an assembly bug for movabs imm, (e/r)ax
This variant of movabs has a special encoding for moving to eax/rax.  The
immediate is stored in a format we haven't run across before, so this commit
adds some support for encoding it properly (and also avoiding trying to encode
operands of type OffsetSource as ModR/M bytes.  This movabs variant uses
OffsetSource to encode the immediate, which is morally equivalent to
ImmediateSource, which also should not be encoded as a ModR/M byte.

We also had to add some cases for encoding OffsetSources