Fix the definition of the fld instruction
The normal cases (of memory references, as opposed to x87 FP stack references)
had a `/mode` constraint that wasn't present in the manual (and caused errors
while re-assembling, because they fixed the mode to the wrong value).

I think that the constraints were added for the x87 FP stack references, where
they are probably actually required to get the right mode value, but they aren't
correct for the memory reference versions.

This fixes #4.