Have instruction encoding fail more gracefully
Before, it would call error in a few cases, which was painful to deal with at
call sites.  Now, use MonadThrow to fail with structured errors, allowing the
caller to discard them (use as a Maybe) or catch purely (with Either).

This is especially important as we can now decode some AVX instructions, but
can't yet encode them.