MonadFail updates for GHC 8.8 compatibility
This modifies some uses of the `ByteReader` interface to use `error` instead of
`fail`.  None of the types implementing `ByteReader` can easily implement `MonadFail` because
they are based on State.  There are already many uses of error around these
updates, so it isn't unprecedented.

We could add some more structure to the error handling of some of these cases,
but it would be a substantial amount of work (or require the use sites to change
to a different base monad).