Fix a bug introduced while adding ghc-8.8 compatibility
This fixes #15.  The first attempt at ghc-8.8 changed uses of `fail` in the
`ByteReader` to use error instead, as it wasn't immediately obvious that all of
the concrete classes implementing it could be instances of MonadFail.  It turns
out that they can be and that we are relying on that for correct behavior.  This
change reverts to using `fail` in a few places.

It would be really nice to make these failures a bit more structured.