Sign-extend immediate operands in `push` (#37)
As noted in #35, immediate operands in `push` should be sign-extended in the
immediate is of a size less than the operand size. In `flexdis86`, we can
encode this by using the appropriate sign-extended `Value` constructors. This
amounts to a couple of minor tweaks in `data/optable.xml`. I have also made
a first attempt at making `mkInstruction` infer whether operand-size overrides
are used, which is necessary in order to make the test suite pass.